requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
csscompressor==0.9.5
docutils==0.19
Flask==2.2.2
Flask-Caching==2.0.2
Flask-CKEditor==0.4.6
Flask-WTF==1.1.1
flask-googlemaps==0.4.1
googlemaps==4.10.0
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mysql-connector==2.2.9
noise==1.2.2
packaging==23.0
perlin-noise==1.12
Pillow==9.4.0
pygame==2.1.2
Pygments==2.14.0
pymdown-extensions==9.9.2
python-dateutil===2.8.1
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14
watchdog==2.2.1
Werkzeug==2.2.2
WTForms==3.0.1
